Crate ckb_jsonrpc_types

Modules

error
jsonrpc errors

id
jsonrpc id field

params
jsonrpc params field

request
jsonrpc request

response
jsonrpc response

version
jsonrpc version field

Structs

Alert

AlertMessage

BannedAddr

Block

BlockEconomicState

BlockIssuance

BlockReward

BlockTemplate

BlockView

Byte32

CellDep

CellInput

CellOutput

CellOutputWithOutPoint

CellTransaction

CellWithStatus

CellbaseTemplate

ChainInfo

DryRunResult

EpochView

EstimateResult

Header

HeaderView

JsonBytes

LiveCell

LocalNode

LockHashCapacity

LockHashIndexState

MinerReward

NodeAddress

OutPoint

PeerState

ProposalShortId

RemoteNode

Script

Transaction

TransactionPoint

TransactionTemplate

TransactionView

TransactionWithStatus

TxPoolInfo

TxStatus

UncleBlock

UncleBlockView

UncleTemplate

Enums

DepType

OutputsValidator

ScriptHashType

Status
Status for transaction

Traits

Deserialize
A data structure that can be deserialized from any data format supported by Serde.

Serialize
A data structure that can be serialized into any data format supported by Serde.

Type Definitions

BlockNumber

Capacity

Cycle

EpochNumber

EpochNumberWithFraction

FeeRate

Timestamp

Uint32

Uint64

Uint128

Version

Derive Macros

Deserialize

Serialize